// src/types/user.ts
export type UserProfile = {
  readonly id: string;
  readonly x_user_id: string;
  readonly x_handle: string;
  readonly x_display_name: string;
  readonly x_avatar_url: string;
  readonly created_at: string;
  readonly last_login_at: string;
};

export type UserSummary = {
  readonly x_handle: string;
  readonly x_display_name: string;
  readonly x_avatar_url: string;
};